Fix DataGrid column autogeneration
By using ObservableCollection<obj> instead of ObservableCollection<ViewModel<obj, obj>> we can force CollectionView.GetItemProperties() to inspect the first item in the collection in order to find the columns to display. If we don't do this, it will instead inspect ViewModel<obj, obj> instead of the derived type we emit at runtime, and fail to find any properties of interest.