Don't create types unnecessarily
1. Cache the created types
2. Don't call getBindings() unnecessarily

Each time getBindings() is invoked, a new array of bindings is created.

These arrays are used as the keys to the types generated from them in the type cache.

The type cache will only ever grow, so I don't think it would be hard to  create a program with a significant memory leak, because there's no absolute restriction to a single invocation of each getBindings() method (see SubBindings sample for an example where it'll be invoked multiple times due to it recursively referencing itself).